+++ src/lib/lisp/liblisp.h
@@ -144,12 +144,44 @@
 #define MALLOCD_HEAP
 #define MALLOCD_STACK
 #define MALLOCD_TAGSTACK
-#define STACK_SIZE          2048
+#define STACK_SIZE          1024
 #define TAGSTACK_SIZE       512
 #define RELOC_TABLE_ENTRIES 256
 #define SKIPPING_SWEEP
 #define PRINT_SHORT_QUOTES
 #define MAX_SYMBOL  (255 - sizeof (symbol))
+#define NO_TIME
+#define NO_DIRECTORY
+#endif
+
+// Apple II enhanced
+#ifdef TARGET_APPLE2ENH
+#define MALLOCD_HEAP
+#define MALLOCD_STACK
+#define MALLOCD_TAGSTACK
+#define STACK_SIZE          1024
+#define TAGSTACK_SIZE       512
+#define RELOC_TABLE_ENTRIES 256
+#define SKIPPING_SWEEP
+#define PRINT_SHORT_QUOTES
+#define MAX_SYMBOL  (255 - sizeof (symbol))
+#define NO_TIME
+#define NO_DIRECTORY
+#endif
+
+// Atari XL
+#ifdef TARGET_ATARIXL
+#define MALLOCD_HEAP
+#define MALLOCD_STACK
+#define MALLOCD_TAGSTACK
+#define STACK_SIZE          1024
+#define TAGSTACK_SIZE       512
+#define RELOC_TABLE_ENTRIES 256
+#define SKIPPING_SWEEP
+#define PRINT_SHORT_QUOTES
+#define MAX_SYMBOL  (255 - sizeof (symbol))
+#define NO_TIME
+#define NO_DIRECTORY
 #endif
 
 // Commodore C128
@@ -166,6 +198,7 @@
 #define SKIPPING_SWEEP
 #define PRINT_SHORT_QUOTES
 #define MAX_SYMBOL  (255 - sizeof (symbol))
+#define NO_TIME
 #endif
 
 // Commodore C16
@@ -213,6 +246,8 @@
 #define SKIPPING_SWEEP
 #define PRINT_SHORT_QUOTES
 #define MAX_SYMBOL  (255 - sizeof (symbol))
+#define NO_TIME
+#define NO_DIRECTORY
 #endif
 
 // Commodore PET
@@ -255,6 +290,8 @@
 #define SKIPPING_SWEEP
 #define PRINT_SHORT_QUOTES
 #define MAX_SYMBOL  (255 - sizeof (symbol))
+#define NO_TIME
+#define NO_DIRECTORY
 #endif
 
 // Commodore VIC-20/VC-20
@@ -289,6 +326,7 @@
 #define SKIPPING_SWEEP
 #define PRINT_SHORT_QUOTES
 #define MAX_SYMBOL  65536
+#define NO_DIRECTORY
 #endif
 
 // Sinclair ZX Spectrum
@@ -301,6 +339,8 @@
 #define RELOC_TABLE_ENTRIES 512
 #define SKIPPING_SWEEP
 #define PRINT_SHORT_QUOTES
+#define NO_TIME
+#define NO_DIRECTORY
 #endif
 
 #if !defined (MALLOCD_HEAP) && !defined (HEAP_SIZE)